git-svn-id: svn+ssh://svn.code.sf.net/p/lemur/code/indri/trunk@301 14020d50-3050-45fe-8e0c-5ed7005ca764

+++ include/indri/QueryResponsePacker.hpp
@@ -50,7 +50,7 @@ public:
         size_t resultsSent = 0;
 
         while( resultList.size() > resultsSent ) {
-          size_t sendChunk = lemur_compat::min<size_t>( resultList.size() - resultsSent, 100 );
+          size_t sendChunk = lemur_compat::min<size_t>( resultList.size() - resultsSent, (size_t) 100 );
 
           for( size_t i=0; i<sendChunk; i++ ) {
             ScoredExtentResult byteSwapped;
